Add files via upload
CHINA'S Step towards Superpower: A Debt trap for Countries worldwide. This is a detailed analysis of the loan sanctioned by China to various countries all around the world from 2009 to 2019 and the Sectors where China invested the most. A detailed study about the Sri Lanka, Pakistan and Ukraine is also there with the beautiful visualisations using bar plots and histograms